Move module Kore.ASTUtils.Substitution to Kore.Proof.Substitution
The module Kore.ASTUtils.Substitution should not be used outside the proof
modules.
