+++ src/main/haskell/kore/src/Kore/Proof/Substitution.hs
@@ -1,5 +1,5 @@
 {-|
-Module      : Kore.ASTUtils.Substitution
+Module      : Kore.Proof.Substitution
 Description : Substitute phi_1 for phi_2, avoiding capture
               In particular this implements axiom 7 in
               the "large" axiom set (Rosu 2017).
@@ -8,11 +8,15 @@ License     : NCSA
 Maintainer  : phillip.harris@runtimeverification.com
 Stability   : experimental
 Portability : portable
+
+This implementation of substitution allows the substitution target to be an
+arbitrary pattern, not just a variable. It should not be used outside
+"Kore.Proof.Proof" and the test suite.
 -}
 
 {-# LANGUAGE AllowAmbiguousTypes #-}
 
-module Kore.ASTUtils.Substitution
+module Kore.Proof.Substitution
     ( subst
     ) where
