+++ src/main/haskell/kore/test/Test/Kore/ASTUtils.hs
@@ -2,8 +2,7 @@
 {-# OPTIONS_GHC -Wno-missing-pattern-synonym-signatures #-}
 
 module Test.Kore.ASTUtils
-    ( test_substitutions
-    , test_sortAgreement
+    ( test_sortAgreement
     , var
     , var_
     , mkSort
@@ -14,35 +13,15 @@ import Test.Tasty
 import Test.Tasty.HUnit
        ( assertEqual, testCase )
 
-import           Control.Lens
-import           Data.Text
-                 ( Text )
-import qualified Data.Text as Text
+import Control.Lens
+import Data.Text
+       ( Text )
 
 import Kore.AST.Lens
        ( resultSort )
 import Kore.AST.Pure
 import Kore.AST.Valid
-import Kore.ASTUtils.Substitution
 import Kore.Step.Pattern
-
-test_substitutions :: TestTree
-test_substitutions = testGroup "Substitutions"
-    [ testCase "subTrivial" $ assertEqual "" subTrivial subTrivialSolution
-    , testCase "subShadow"  $ assertEqual "" subShadow subShadowSolution
-    , testCase "subAlphaRename1" $
-          assertEqual ""
-              subAlphaRename1
-              subAlphaRename1Solution
-    , testCase "subAlphaRename2" $
-          assertEqual ""
-              (subAlphaRename2 ^? inPath [0])
-              (Just $ mkVar $ var "b")
-    , testCase "subTermForTerm" $
-          assertEqual ""
-              subTermForTerm
-              subTermForTermSolution
-    ]
 
 test_sortAgreement :: TestTree
 test_sortAgreement = testGroup "Sort agreement"
@@ -73,48 +52,7 @@ test_sortAgreement = testGroup "Sort agreement"
             (Just (predicateSort :: Sort Object))
     , testGroup "sortAgreementManySimplePatterns"
         sortAgreementManySimplePatterns
-    , testGetSetIdentity 5
     ]
-
-subTrivial :: CommonStepPattern Object
-subTrivial =
-    subst (mkVar $ var "a") (mkVar $ var "b") $
-    mkExists (var "p") (mkVar $ var "a")
-
-subTrivialSolution :: CommonStepPattern Object
-subTrivialSolution = mkExists (var "p") (mkVar $ var "b")
-
-subShadow :: CommonStepPattern Object
-subShadow =
-    subst (mkVar $ var "a") (mkVar $ var "b") $
-    mkExists (var "a") (mkVar $ var "q")
-
-subShadowSolution :: CommonStepPattern Object
-subShadowSolution =
-    mkExists (var "a") (mkVar $ var "q")
-
-subAlphaRename1 :: CommonStepPattern Object
-subAlphaRename1 =
-    subst (mkVar $ var "a") (mkVar $ var "b") $
-    mkExists (var "b") (mkVar $ var "q")
-
-subAlphaRename1Solution :: CommonStepPattern Object
-subAlphaRename1Solution =
-    mkExists (var "b0") (mkVar $ var "q")
-
-subAlphaRename2 :: CommonStepPattern Object
-subAlphaRename2 =
-    subst (mkVar $ var "a") (mkVar $ var "b") $
-    mkExists (var "b") (mkVar $ var "a")
-
-subTermForTerm :: CommonStepPattern Object
-subTermForTerm =
-    subst (mkOr mkTop_ mkBottom_) (mkAnd mkTop_ mkBottom_) $
-    mkImplies (mkOr mkTop_ mkBottom_) mkTop_
-
-subTermForTermSolution :: CommonStepPattern Object
-subTermForTermSolution =
-    mkImplies (mkAnd mkTop_ mkBottom_) mkTop_
 
 -- subAlphaRename2Solution :: CommonStepPattern Object
 -- subAlphaRename2Solution = dummyEnvironment @Object $
@@ -175,38 +113,6 @@ sortAgreementManySimplePatterns = do
             (getSort shouldHavepredicateSortTwoArgs)
             predicateSort
     assert1 ++ assert2 ++ assert3 ++ assert4
-
-substitutionGetSetIdentity a b pat =
-  assertEqual ""
-  (subst b a pat)
-  (subst b a $ subst a b pat)
-
-generatePatterns :: Int -> [CommonStepPattern Object]
-generatePatterns size = genBinaryPatterns size ++ genUnaryPatterns size
-genBinaryPatterns :: Int -> [CommonStepPattern Object]
-genBinaryPatterns 0 = []
-genBinaryPatterns size = do
-  sa <- [1..size-1]
-  let sb = size - sa
-  a <- generatePatterns sa
-  b <- generatePatterns sb
-  [mkAnd a b, mkOr a b, mkImplies a b, mkIff a b, mkRewrites a b]
-genUnaryPatterns :: Int -> [CommonStepPattern Object]
-genUnaryPatterns 0 = []
-genUnaryPatterns 1 = [mkVar $ var_ "x" "X"]
-genUnaryPatterns size = do
-  a <- generatePatterns (size - 1)
-  [mkNot a, mkNext a, mkForall (var $ Text.pack $ show size) a]
-
---FIXME: Make a proper Tasty generator instead
-testGetSetIdentity
-  :: Int
-  -> TestTree
-testGetSetIdentity size = testGroup "getSetIdent" $ do
-  a <- generatePatterns (size `div` 3)
-  b <- generatePatterns (size `div` 3)
-  pat <- generatePatterns size
-  return $ testCase "" $ substitutionGetSetIdentity a b pat
 
 var :: MetaOrObject level => Text -> Variable level
 var x = Variable (noLocationId x) (mkSort "S")
